1.Create a CreateAccount class that used for user and admin accounts registration 2.Minor changes in SignUpController.java and AdminPageController.java 3.Set role to 0 when user creates a user account and set role to 1 when user creates an admin account
